# optimizer: Julia-level escape analysis