… base/compiler/ssair/EscapeAnalysis/EscapeAnalysis.jl
… doc/src/devdocs/EscapeAnalysis.md
… src/gf.c
… src/jltypes.c
… test/choosetests.jl
… test/compiler/EscapeAnalysis/EAUtils.jl
… test/compiler/EscapeAnalysis/interprocedural.jl
… test/compiler/EscapeAnalysis/local.jl
… test/compiler/EscapeAnalysis/setup.jl