@nanosoldier 2022-01-14
[Your package evaluation job](https://github.com/JuliaLang/julia/pull/43800#issuecomment-1012502417) has completed - possible new issues were detected. A full report can be found [here](https://s3.amazonaws.com/julialang-reports/nanosoldier/pkgeval/by_hash/faf954b_vs_89f2332/report.html).