@JeffBezanson 2022-01-21
I get this during build:
```
Generating REPL precompile statements... 36/36
Executing precompile statements... 121/1349[EA] bad EA: Array{Any, (2,)}[
  LibGit2.var"#148#149"{LibGit2.GitRepo, String, String},
  LibGit2.GitReference] at 134
[EA] bad EA: Array{Any, (2,)}[
  LibGit2.var"#148#149"{LibGit2.GitRepo, String, String},
  LibGit2.GitReference] at 162
```