@aviatesk 2022-01-22
> I get this during build:
> 
> ```
> Generating REPL precompile statements... 36/36
> Executing precompile statements... 121/1349[EA] bad EA: Array{Any, (2,)}[
>   LibGit2.var"#148#149"{LibGit2.GitRepo, String, String},
>   LibGit2.GitReference] at 134
> [EA] bad EA: Array{Any, (2,)}[
>   LibGit2.var"#148#149"{LibGit2.GitRepo, String, String},
>   LibGit2.GitReference] at 162
> ```

This is because this PR uses the simple `is_load_forwardable` to check if EA covers all load-forwardable structs that the currently existing `sroa_pass!` can find, but that simple `is_load_forwardable` check doesn't account for ccall-preserve uses, which `sroa_pass!` can sometimes rewrite with forwarded loads by some extra efforts. #43888 will rewrite mutable SROA pass based on EA, and it is able to account for such ccall-preserve uses and we won't miss this optimization opportunity.