@aviatesk 2022-01-25
From an online discussion:

TODOs before merging:
- [ ] remove the suspicious warning
- [ ] move EAUtils to the test folder and leave a comment for generalization in the future
- [ ] fix correctness of `aliasset`
- [ ] fix correctness of `AliasInfo` (as we see in #43909)
- [ ] change the `ThrownEscape` sentinel
- [ ] setup a proper cache infrastructure (integration with `CodeInstance`)
- [ ] check the performance of EA if it runs before inlining (for IPO analysis)

And we will merge #43888 and #43909 separately after finishing up this PR.