@aviatesk 2022-02-03
ok, let me update the latency issue: I re-measured build time, [`CCProfiler.jl`](https://gist.github.com/aviatesk/54ffa6e99d77e7bd8824e3616961de7f), and TTFP on amdci2 with the latest commits of EA-related PRs.
Comparison are made for these setups:
- `master` (c10dac100a)
- this PR (6c740f83ce): run `IPO EA` before inlining for IPO cache generation
- #43888 (9f86d67f90): run `IPO EA` before inlining for IPO cache generation and also `Local EA` for SROA of mutables

### Build log

> `master` (c10dac100a)
```
Sysimage built. Summary:
Total ───────  59.874610 seconds 
Base: ───────  24.087176 seconds 40.2294%
Stdlibs: ────  35.786103 seconds 59.7684%
    JULIA usr/lib/julia/sys-o.a
Generating REPL precompile statements... 36/36
Executing precompile statements... 1303/1332
Precompilation complete. Summary:
Total ─────── 111.193477 seconds
Generation ──  83.259207 seconds 74.8778%
Execution ───  27.934269 seconds 25.1222%
```

> this PR (6c740f83ce)
```
Sysimage built. Summary:
Total ───────  60.866123 seconds 
Base: ───────  24.391461 seconds 40.074%
Stdlibs: ────  36.473287 seconds 59.9238%
    JULIA usr/lib/julia/sys-o.a
Generating REPL precompile statements... 36/36
Executing precompile statements... 1310/1339
Precompilation complete. Summary:
Total ─────── 114.789126 seconds
Generation ──  87.021880 seconds 74.512%
Execution ───  29.767246 seconds 25.488%
```

> #43888 (9f86d67f90)
```
Sysimage built. Summary:
Total ───────  60.953121 seconds 
Base: ───────  24.508623 seconds 40.209%
Stdlibs: ────  36.443151 seconds 59.7888%
    JULIA usr/lib/julia/sys-o.a
Generating REPL precompile statements... 36/36
Executing precompile statements... 1312/1341
Precompilation complete. Summary:
Total ─────── 118.214727 seconds
Generation ──  87.707580 seconds 74.1934%
Execution ───  30.507147 seconds 25.8066%
```

### TTFP

> `master` (c10dac100a)
```
❯ ./usr/bin/julia -e "@time using Plots; @time plot(rand(10,3));" 
  3.705093 seconds (8.07 M allocations: 549.432 MiB, 3.77% gc time, 20.85% compilation time)
  4.631320 seconds (14.23 M allocations: 809.206 MiB, 11.30% gc time, 99.84% compilation time)
```

> this PR (6c740f83ce)
```
❯ ./usr/bin/julia -e "@time using Plots; @time plot(rand(10,3));" 
  3.872518 seconds (8.11 M allocations: 552.244 MiB, 5.03% gc time, 21.79% compilation time)
  4.659757 seconds (14.67 M allocations: 839.512 MiB, 9.65% gc time, 99.84% compilation time)
```

> #43888 (9f86d67f90)
```
❯ ./usr/bin/julia -e "@time using Plots; @time plot(rand(10,3));" 
  3.828619 seconds (8.01 M allocations: 540.189 MiB, 8.72% gc time, 20.67% compilation time)
  4.604284 seconds (14.77 M allocations: 849.715 MiB, 7.07% gc time, 99.81% compilation time)
```

### [`CCProfiler.jl`](https://gist.github.com/aviatesk/54ffa6e99d77e7bd8824e3616961de7f)

> `master` (c10dac100a)
```
❯ ./usr/bin/julia ../CCProfiler.jl 
[ Info: Base.return_types(println, (QuoteNode,), CCProfiler())
  0.812259 seconds (5.33 M allocations: 280.519 MiB, 34.38% gc time, 100.00% compilation time)
[ Info: Base.return_types(println, (QuoteNode,), CCProfiler())
  0.587490 seconds (5.32 M allocations: 279.846 MiB, 13.32% gc time, 100.00% compilation time)
[ Info: Base.return_types(println, (QuoteNode,), CCProfiler())
  0.756882 seconds (5.32 M allocations: 279.846 MiB, 33.68% gc time, 100.00% compilation time)
[ Info: Base.return_types(sin, (Int,), CCProfiler())
  0.034916 seconds (454.73 k allocations: 24.779 MiB, 99.95% compilation time)
[ Info: Base.return_types(Base.init_stdio, (Ptr{Cvoid},), CCProfiler())
  3.017604 seconds (21.84 M allocations: 1.109 GiB, 21.01% gc time, 100.00% compilation time)
[ Info: Base.return_types(CC.construct_ssa!, (Core.CodeInfo, CC.IRCode, CC.DomTree, Vector{CC.SlotInfo}, Vector{Any}), CCProfiler())
  0.534155 seconds (4.81 M allocations: 252.316 MiB, 22.37% gc time, 100.00% compilation time)
[ Info: Base.return_types(CC.domsort_ssa!, (CC.IRCode, CC.DomTree), CCProfiler())
  0.182586 seconds (1.53 M allocations: 81.688 MiB, 30.52% gc time, 99.99% compilation time)
```

> this PR (6c740f83ce)
```
❯ ./usr/bin/julia ../CCProfiler.jl 
[ Info: Base.return_types(println, (QuoteNode,), CCProfiler())
  1.071022 seconds (5.65 M allocations: 297.476 MiB, 22.89% gc time, 20.44% compilation time)
[ Info: Base.return_types(println, (QuoteNode,), CCProfiler())
  0.629850 seconds (5.41 M allocations: 285.193 MiB, 15.54% gc time, 100.00% compilation time)
[ Info: Base.return_types(println, (QuoteNode,), CCProfiler())
  0.769553 seconds (5.41 M allocations: 285.193 MiB, 31.26% gc time, 100.00% compilation time)
[ Info: Base.return_types(sin, (Int,), CCProfiler())
  0.051713 seconds (455.80 k allocations: 24.825 MiB, 30.19% gc time, 99.97% compilation time)
[ Info: Base.return_types(Base.init_stdio, (Ptr{Cvoid},), CCProfiler())
  3.348238 seconds (22.42 M allocations: 1.144 GiB, 24.96% gc time, 100.00% compilation time)
[ Info: Base.return_types(CC.construct_ssa!, (Core.CodeInfo, CC.IRCode, CC.DomTree, Vector{CC.SlotInfo}, Vector{Any}), CCProfiler())
  0.584729 seconds (5.06 M allocations: 275.882 MiB, 20.15% gc time, 100.00% compilation time)
[ Info: Base.return_types(CC.domsort_ssa!, (CC.IRCode, CC.DomTree), CCProfiler())
  0.173689 seconds (1.59 M allocations: 85.794 MiB, 22.17% gc time, 99.99% compilation time)
```

> #43888 (9f86d67f90)
```
❯ ./usr/bin/julia ../CCProfiler.jl 
[ Info: Base.return_types(println, (QuoteNode,), CCProfiler())
  0.982920 seconds (5.72 M allocations: 306.689 MiB, 12.03% gc time, 36.20% compilation time)
[ Info: Base.return_types(println, (QuoteNode,), CCProfiler())
  0.786400 seconds (5.43 M allocations: 287.642 MiB, 32.67% gc time, 100.00% compilation time)
[ Info: Base.return_types(println, (QuoteNode,), CCProfiler())
  0.766854 seconds (5.43 M allocations: 287.642 MiB, 31.06% gc time, 100.00% compilation time)
[ Info: Base.return_types(sin, (Int,), CCProfiler())
  0.039858 seconds (455.18 k allocations: 24.938 MiB, 99.93% compilation time)
[ Info: Base.return_types(Base.init_stdio, (Ptr{Cvoid},), CCProfiler())
  3.463343 seconds (22.66 M allocations: 1.170 GiB, 24.95% gc time, 100.00% compilation time)
[ Info: Base.return_types(CC.construct_ssa!, (Core.CodeInfo, CC.IRCode, CC.DomTree, Vector{CC.SlotInfo}, Vector{Any}), CCProfiler())
  0.676202 seconds (5.28 M allocations: 303.039 MiB, 20.28% gc time, 100.00% compilation time)
[ Info: Base.return_types(CC.domsort_ssa!, (CC.IRCode, CC.DomTree), CCProfiler())
  0.174834 seconds (1.60 M allocations: 86.372 MiB, 20.62% gc time, 99.99% compilation time)
```

---

So now I think the numbers are "acceptable" (the increased allocations are about in ~105% range).
Especially ad6bfa3813 was very effective to mitigate the additional compilation cost by introducing an heuristic to avoid running `IPO EA` when there are no "interesting" arguments in terms of their escapability.
Also, if we don't land the `ImmutableArray` PR in v1.8, we can even disable `IPO EA` entirely before releasing a stable version since the EA-based SROAs won't much benefit from IPO escape information, and then the additional compilation costs would be negligible.